requirements.txt
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0